requirements.txt
streamlit
requests
python-dotenv
pillows